Add Windows support for WriteSubnetFile via MoveFileEx
